fixed username and password
Changed the curl option for username and password from -U to -u. -U ist for the proxy username and password. -u ist for the http username and password.

Removed quotes arround username and password: They were treated as pert of the username and password.

Added curlopt to the restore curl commands to use username and password there as well.